cloud_file_uploader/requirements.txt
boto3
google-cloud-storage
pytest
pytest-cov
